build: Replace `DeleteArtifacts` target with metadata and properties
Instead of manually deleting references and the .deps.json file in the output directory, the `Private` metadata can be set to `false` in the `Reference` items to skip copying those references to the output directory. Furthermore, the generation of the .deps.json file can be disabled with the `GenerateDependencyFile` property.